requirements.txt
discord.py
requests
python-decouple
pyparsing
